Flask_WebAPI/requirements.txt
flask==1.1.2
flask_wtf==0.14.3
Card_Validator==1.0.0
python-dotenv==0.15.0